+++ packages/ui/src/internal/appearance.ts
@@ -395,10 +395,10 @@ export type ElementsConfig = {
   organizationProfileSecuritySsoBypassSearchInput: WithOptions;
   organizationProfileSecuritySsoBypassSearchInputIcon: WithOptions;
   organizationProfileSecuritySsoBypassAddButton: WithOptions;
-  organizationProfileSecuritySsoBypassMemberSearchInput: WithOptions;
-  organizationProfileSecuritySsoBypassMemberSearchInputIcon: WithOptions;
-  organizationProfileSecuritySsoBypassMemberOptions: WithOptions;
-  organizationProfileSecuritySsoBypassMemberOption: WithOptions;
+  organizationProfileSecuritySsoBypassEmailInput: WithOptions;
+  organizationProfileSecuritySsoBypassRoleWarning: WithOptions;
+  organizationProfileSecuritySsoBypassFailure: WithOptions;
+  organizationProfileSecuritySsoBypassBulkResult: WithOptions;
 
   organizationListPreviewItems: WithOptions;
   organizationListPreviewItem: WithOptions;